Add ModelManager test code and fix bug
More test code for ModelManager are added.

Fixed a bug in Trie that when removing the only single word in the
Trie, it gives a NullPointerException.
The logic does not recognize the root node and hence when trying to
get the parent of root node, which is null, gives an exception when
invoking method from that null parent.